test: split guardrails remote runtime coverage
